Completed Registration and Authentication functions. Now I will work on the library.
